fix(core): pin unplugin-ast and move to metaPlugin
