+++ packages/core/src/index.ts
@@ -1,7 +1,4 @@
 import { joinURL } from 'ufo'
-import AST from 'unplugin-ast/vite'
-import type { Transformer } from 'unplugin-ast'
-import type { CallExpression } from '@babel/types'
 
 import { libraryCollections as libraryCollectionsConfig } from '@compodium/examples'
 import { collectionsPlugin } from './plugins/collections'
@@ -14,19 +11,6 @@ import { iconifyPlugin } from './plugins/iconify'
 import type { Collection, PluginConfig, PluginOptions } from './types'
 
 export * from './types'
-
-const RemoveFunction = (
-  functionNames: string[]
-): Transformer<CallExpression> => ({
-  onNode: node =>
-    node.type === 'CallExpression'
-    && node.callee.type === 'Identifier'
-    && functionNames.includes(node.callee.name),
-
-  transform() {
-    return null
-  }
-})
 
 export const compodium = /* #__PURE__ */ (options: PluginOptions) => {
   const exampleDir = {
@@ -78,10 +62,6 @@ export const compodium = /* #__PURE__ */ (options: PluginOptions) => {
     devtoolsPlugin(config),
     examplePlugin(config),
     iconifyPlugin(config),
-    colorsPlugin(config),
-    AST({
-      include: [/\.[jt]sx?$/],
-      transformer: [RemoveFunction(['extendCompodiumMeta'])]
-    })
+    colorsPlugin(config)
   ]
 }
